// src/enums/pipedrive.enum.ts
export enum PipeDriveDealsField {
    FURNISHING_LEVEL = '',
    LIST_OF_AMENITIES_AVAILABLE = '',
    LIST_OF_FURNITURES_AVAILABLE = '',
    NUMBER_OF_BATHROOMS = '',
    NUMBER_OF_BEDROOMS = '',
    NUMBER_OF_PARKING_LOTS = '',
    PHOTOGRAPH_LINK = '',
    PROPERTY_ADDRESS = '',
    PROPERTY_TYPE = '',
    RENTAL_AVAILABLE_DATE = '',
    MONTHLY_RENT = '',
    REQUIRED_DEPOSIT_AMOUNT = '',
    TENANCY_PERIOD = '',
    TYPE_OF_UNIT = '',
    REQUESTED_PHOTOGRAPHY_DATE = '',
    REQUESTED_PHOTOGRAPHY_TIME = '',
    NOTES = '',
    APPOINTMENT_LINK = '',
    REQUESTED_VIEWING_DATE = '',
    REQUESTED_VIEWING_TIME = '',
    SOURCE_OF_LEADS = '',
    LOOKING_FOR_PROPERTY_TYPE = '',
    PREFERRED_LOCATION = '',
    PREFERRED_RESIDENCES = '',
    PHONE = '',
    PAX = '',
    SPECIAL_REQUEST = '',
    NATIONALITY = '',
    OCCUPATION = '',
    COMPANY_UNIVERSITY = '',
    FURNISHING = '',
    MOVE_IN_DATE = '',
    BUDGET_MAX = '',
    VIEWING_CUSTOMER_DETAILS = '',
}

export enum PipeDrivePersonField {
    INSTAHOME_ID = '',
    REN_NUMBER = '',
    DATE_REGISTERED = '',
}