// src/enums/location.enum.ts
export enum PropertyGroupType {
    Landed = 'landed',
    Highrise = 'highrise',
}

export enum LocationLevel {
    Country = 'Country',
    State = 'State',
    Area = 'Area',
    SubArea = 'SubArea',
    Property = 'Property',
}

export enum LocationPropertyType {
    Condominium = 'condominium',
    Apartment = 'apartment',
    Flat = 'flat',
    ServicedResidence = 'serviced_residence',
    Bungalow = 'bungalow',
    TerraceHouse = 'terrace_house',
    SemiDetached = 'semi_detached',
    Townhouse = 'townhouse',
}

export enum LocationLevels {
    Country = 1,
    State = 2,
    Area = 3,
    SubArea = 4,
}